tasks views move
