+++ backend/api/urls.py
@@ -1,12 +1,12 @@
 from django.urls import include, path
 from drf_spectacular.views import SpectacularAPIView, SpectacularSwaggerView
 from rest_framework.routers import DefaultRouter
+from tasks.views import TaskViewSet
 
 from .views import (
     CustomUserViewSet,
     ProgressUserAndDepartmentViewSet,
     ShortUserProfileViewSet,
-    TaskViewSet,
     UserNotificationsViewSet,
 )
